Add byte util

+++ sherpa/csrc/context-graph.cc
@@ -25,9 +25,9 @@
 namespace sherpa {
 void ContextGraph::Build(
     const std::vector<std::vector<int32_t>> &token_ids) const {
-  for (int32_t i = 0; i < token_ids.size(); ++i) {
+  for (size_t i = 0; i < token_ids.size(); ++i) {
     auto node = root_.get();
-    for (int32_t j = 0; j < token_ids[i].size(); ++j) {
+    for (size_t j = 0; j < token_ids[i].size(); ++j) {
       int32_t token = token_ids[i][j];
       if (0 == node->next.count(token)) {
         bool is_end = j == token_ids[i].size() - 1;
